Implement remove component and return/free empty chunks
